blade: update sensors
Blade does not support both light and proximity sensor at the same time. If one is started while the other is active the one that is running will be
stopped, but Android will still think it is active. The lib currently correctly restarts the light sensor when prox is stop if the light sensor should
be running. This patch stops the light sensor from being started if the prox sensor is currently turned on which was causing problems during calls
if auto-brightness was turned on.

Change-Id: I470d440a808f8cb59a6b050f83239c9efa0c34f3