blade: Fix video recording (part 1/2)
This should fix video recording on the Blade.

This is the first part of the commit. The second part is at the propietary side and a pull request can be found at https://github.com/koush/proprietary_vendor_zte/pull/1 which fixes this issue and adds the ZTE Skate as well.

Change-Id: I95810d23fb581428adee9bcff23af89d1be9ab2d